Streamlit/requirements.txt
streamlit
opencv-python
cvzone
numpy
Pillow
google-generativeai
streamlit-option-menu
